@@ -5,20 +5,26 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import kotlinx.android.synthetic.main.fragment_step.view.*
+import com.mediapicker.sample.databinding.FragmentStepBinding
 
-class StepFragment : Fragment(){
+class StepFragment : Fragment() {
 
-        override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
-            return inflater.inflate(R.layout.fragment_step,container,false)
-        }
+    private lateinit var binding: FragmentStepBinding
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        binding = FragmentStepBinding.inflate(layoutInflater)
+        return binding.root
+    }
 
-        override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-            super.onViewCreated(view, savedInstanceState)
-            view.actionButton.setOnClickListener {
-                if(activity is MainActivity){
-                    (activity as MainActivity).jumpToGallery()
-                }
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        binding.actionButton.setOnClickListener {
+            if (activity is MainActivity) {
+                (activity as MainActivity).jumpToGallery()
             }
         }
     }
+}
