@@ -11,17 +11,25 @@ import com.bumptech.glide.Glide
 import com.bumptech.glide.request.RequestOptions
 import com.mediapicker.gallery.Gallery
 import com.mediapicker.gallery.R
+import com.mediapicker.gallery.databinding.OssItemFolderSelectionBinding
 import com.mediapicker.gallery.domain.contract.OnItemClickListener
 import com.mediapicker.gallery.domain.entity.PhotoAlbum
 import com.mediapicker.gallery.domain.entity.PhotoFile
-import kotlinx.android.synthetic.main.oss_item_folder_selection.view.*
 import java.io.File
 
-open class GalleryFolderAdapter constructor(val context: Context, var listOfFolders: List<PhotoAlbum>, val onItemClickListener: OnItemClickListener<PhotoAlbum>?=null) : RecyclerView.Adapter<RecyclerView.ViewHolder>() {
+open class GalleryFolderAdapter constructor(
+    val context: Context,
+    var listOfFolders: List<PhotoAlbum>,
+    val onItemClickListener: OnItemClickListener<PhotoAlbum>? = null
+) : RecyclerView.Adapter<RecyclerView.ViewHolder>() {
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): RecyclerView.ViewHolder {
-        val view = LayoutInflater.from(parent.context).inflate(R.layout.oss_item_folder_selection, parent, false)
-        return FolderViewHolder(view)
+        val binding = OssItemFolderSelectionBinding.inflate(
+            LayoutInflater.from(parent.context),
+            parent,
+            false
+        )
+        return FolderViewHolder(binding)
     }
 
     override fun getItemCount() = listOfFolders.size
@@ -29,15 +37,18 @@ open class GalleryFolderAdapter constructor(val context: Context, var listOfFold
 
     override fun onBindViewHolder(viewHolder: RecyclerView.ViewHolder, position: Int) {
         val vH = viewHolder as FolderViewHolder
-        vH.root.setOnClickListener { v -> onItemClickListener?.onListItemClick(listOfFolders[position]) }
-        vH.root.folderName.text = listOfFolders[position].name
-        vH.root.folderName.isAllCaps = Gallery.galleryConfig.textAllCaps
+        vH.binding.root.setOnClickListener { v -> onItemClickListener?.onListItemClick(listOfFolders[position]) }
+        vH.binding.folderName.text = listOfFolders[position].name
+        vH.binding.folderName.isAllCaps = Gallery.galleryConfig.textAllCaps
         val album = listOfFolders[position]
         if (album.hasPhotos()) {
-            vH.root.backgroundImage.visibility = View.VISIBLE
-            loadImageIntoView(album.firstPhoto as PhotoFile, vH.root.backgroundImage)  //todo shalini: change the type inference
+            vH.binding.backgroundImage.visibility = View.VISIBLE
+            loadImageIntoView(
+                album.firstPhoto as PhotoFile,
+                vH.binding.backgroundImage
+            )  //todo shalini: change the type inference
         } else {
-            vH.root.backgroundImage.visibility = View.GONE
+            vH.binding.backgroundImage.visibility = View.GONE
         }
     }
 
@@ -45,12 +56,14 @@ open class GalleryFolderAdapter constructor(val context: Context, var listOfFold
         val options = RequestOptions()
         if (photo.isAlreadyUploaded) {
             photo.path?.let {
-                Glide.with(imageView.context).load(photo.fullPhotoUrl).apply(options).into(imageView)
+                Glide.with(imageView.context).load(photo.fullPhotoUrl).apply(options)
+                    .into(imageView)
             }
         } else {
             if (photo.existsPhoto()) {
                 photo.path?.let {
-                    Glide.with(imageView.context).load(Uri.fromFile(File(it))).apply(options).into(imageView)
+                    Glide.with(imageView.context).load(Uri.fromFile(File(it))).apply(options)
+                        .into(imageView)
                 }
             } else {
                 photo.toString()
@@ -61,9 +74,10 @@ open class GalleryFolderAdapter constructor(val context: Context, var listOfFold
 }
 
 
-internal class FolderViewHolder(var root: View) : RecyclerView.ViewHolder(root) {
+internal class FolderViewHolder(var binding: OssItemFolderSelectionBinding) :
+    RecyclerView.ViewHolder(binding.root) {
     init {
-        root.img.setImageResource(R.drawable.oss_media_ic_folder_icon)
+        binding.img.setImageResource(R.drawable.oss_media_ic_folder_icon)
     }
 }
 
