@@ -7,6 +7,7 @@ import android.media.ExifInterface
 import android.media.ExifInterface.ORIENTATION_NORMAL
 import android.os.Build
 import android.util.AttributeSet
+import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.FrameLayout
@@ -17,9 +18,8 @@ import androidx.viewpager.widget.ViewPager
 import com.bumptech.glide.Glide
 import com.github.chrisbanes.photoview.PhotoView
 import com.github.chrisbanes.photoview.PhotoViewAttacher
-import com.mediapicker.gallery.R
+import com.mediapicker.gallery.databinding.OssMediaGalleryPagerViewBinding
 import com.mediapicker.gallery.domain.entity.MediaGalleryEntity
-import kotlinx.android.synthetic.main.oss_media_gallery_pager_view.view.*
 import java.io.ByteArrayOutputStream
 import java.io.IOException
 import java.util.*
@@ -36,6 +36,8 @@ open class MediaGalleryPagerView@JvmOverloads constructor(
     private var pinchPanZoomEnabled = false
     private var isGallery = false
     private var mediaChangeListener: MediaChangeListener? = null
+    private var _binding: OssMediaGalleryPagerViewBinding? = null
+    private val binding get() = _binding!!
 
     private val pageChangeListener: ViewPager.OnPageChangeListener = object :
         ViewPager.OnPageChangeListener {
@@ -124,44 +126,44 @@ open class MediaGalleryPagerView@JvmOverloads constructor(
     }
 
     private fun loadDataBasedOnPosition(position: Int) {
-        imageLabel.text = ""
+        binding.imageLabel.text = ""
         setPhotoCount(position + 1, mediaList.size)
         setNavigationIconVisibilityBasedOnPosition(position)
     }
 
     private fun setNavigationIconVisibilityBasedOnPosition(position: Int) {
         if(mediaList.size == 0) {
-            leftArrow.visibility = View.GONE
-            rightArrow.visibility = View.GONE
+            binding.leftArrow.visibility = View.GONE
+            binding.rightArrow.visibility = View.GONE
             return
         }
         if(position == 0) {
-            leftArrow.visibility = View.GONE
+            binding.leftArrow.visibility = View.GONE
         } else {
-            leftArrow.visibility = View.VISIBLE
+            binding.leftArrow.visibility = View.VISIBLE
         }
 
         if(position == mediaList.size - 1) {
-            rightArrow.visibility = View.GONE
+            binding.rightArrow.visibility = View.GONE
         } else {
-            rightArrow.visibility = View.VISIBLE
+            binding.rightArrow.visibility = View.VISIBLE
         }
     }
 
     private fun setPhotoCount(currentImage: Int, size: Int) {
-        if (photoCount != null) {
+        if (binding.photoCount != null) {
             if (size > 1) {
-                photoCount.visibility = View.VISIBLE
-                photoCount.text = String.format(Locale.ENGLISH, " %1\$d / %2\$d ", currentImage, size)
+                binding.photoCount.visibility = View.VISIBLE
+                binding.photoCount.text = String.format(Locale.ENGLISH, " %1\$d / %2\$d ", currentImage, size)
             } else {
-                photoCount.visibility = View.GONE
+                binding.photoCount.visibility = View.GONE
             }
         }
     }
 
     private fun resolvePlaceHolder() {
         if (isGallery) {
-            image_placeholder!!.visibility = View.GONE
+            binding.imagePlaceholder.visibility = View.GONE
         }
     }
 
@@ -186,11 +188,11 @@ open class MediaGalleryPagerView@JvmOverloads constructor(
     }
 
     fun setSelectedPhoto(position: Int) {
-        itemImages!!.currentItem = adapter.resolveItemPosition(position)
+        binding.itemImages.currentItem = adapter.resolveItemPosition(position)
     }
 
     val currentItem: Int
-        get() = adapter.resolveItemPosition(itemImages!!.currentItem)
+        get() = adapter.resolveItemPosition(binding.itemImages.currentItem)
 
     val mediaListSize: Int
         get() = mediaList.size
@@ -248,7 +250,7 @@ open class MediaGalleryPagerView@JvmOverloads constructor(
 
         private fun loadImage(media: MediaGalleryEntity) {
             val imageUrl: String = media.path!!
-            image_progress!!.visibility = View.VISIBLE
+            binding.imageProgress.visibility = View.VISIBLE
             img!!.setOnClickListener(onItemClickListener)
             if(media.isLocalImage) {
                 img!!.tag = media
@@ -303,7 +305,7 @@ open class MediaGalleryPagerView@JvmOverloads constructor(
 
         override fun destroyItem(container: ViewGroup, position: Int, `object`: Any) {
             var view = `object` as View?
-            itemImages.removeView(view)
+            binding.itemImages.removeView(view)
         }
 
         override fun isViewFromObject(view: View, `object`: Any): Boolean {
@@ -316,38 +318,39 @@ open class MediaGalleryPagerView@JvmOverloads constructor(
     }
 
     init {
-        View.inflate(context, R.layout.oss_media_gallery_pager_view, this)
+        _binding = OssMediaGalleryPagerViewBinding.inflate(LayoutInflater.from(context), this)
+
         adapter = ImagePageAdapter()
-        itemImages!!.addOnPageChangeListener(pageChangeListener)
-        itemImages!!.adapter = adapter
-        leftArrow.setOnClickListener(leftArrowClickListener)
-        rightArrow.setOnClickListener(rightArrowClickListener)
+        binding.itemImages.addOnPageChangeListener(pageChangeListener)
+        binding.itemImages.adapter = adapter
+        binding.leftArrow.setOnClickListener(leftArrowClickListener)
+        binding.rightArrow.setOnClickListener(rightArrowClickListener)
         setViewBasedOnMediaList()
     }
 
     private fun setViewBasedOnMediaList() {
         navigationArrowVisibility()
         if(mediaList.size > 0) {
-            defaultContainer.visibility = View.GONE
-            blackBackground.visibility = View.VISIBLE
-            galleryPagerGradient.visibility = View.GONE
-            itemImages.visibility = View.VISIBLE
+            binding.defaultContainer.visibility = View.GONE
+            binding.blackBackground.visibility = View.VISIBLE
+            binding.galleryPagerGradient.visibility = View.GONE
+            binding.itemImages.visibility = View.VISIBLE
         } else {
-            defaultContainer.visibility = View.VISIBLE
-            blackBackground.visibility = View.GONE
-            galleryPagerGradient.visibility = View.GONE
-            image_progress.visibility = View.GONE
-            itemImages.visibility = View.GONE
+            binding.defaultContainer.visibility = View.VISIBLE
+            binding.blackBackground.visibility = View.GONE
+            binding.galleryPagerGradient.visibility = View.GONE
+            binding.imageProgress.visibility = View.GONE
+            binding.itemImages.visibility = View.GONE
         }
     }
 
     private fun navigationArrowVisibility() {
         if(mediaList.size > 1) {
-            leftArrow.visibility = View.VISIBLE
-            rightArrow.visibility = View.VISIBLE
+            binding.leftArrow.visibility = View.VISIBLE
+            binding.rightArrow.visibility = View.VISIBLE
         } else {
-            leftArrow.visibility = View.GONE
-            rightArrow.visibility = View.GONE
+            binding.leftArrow.visibility = View.GONE
+            binding.rightArrow.visibility = View.GONE
         }
     }
 }
