@@ -9,24 +9,27 @@ import androidx.annotation.CallSuper
 import androidx.annotation.LayoutRes
 import androidx.fragment.app.Fragment
 import com.mediapicker.gallery.R
+import com.mediapicker.gallery.databinding.OssFragmentBaseBinding
 import com.mediapicker.gallery.domain.entity.PhotoFile
 import com.mediapicker.gallery.presentation.viewmodels.VideoFile
-import kotlinx.android.synthetic.main.oss_custom_toolbar.*
-import kotlinx.android.synthetic.main.oss_fragment_base.*
-import kotlinx.android.synthetic.main.oss_fragment_base.view.*
 
 open abstract class BaseFragment : Fragment() {
 
-    companion object{
+    companion object {
         const val EXTRA_SELECTED_PHOTOS = "selected_photos"
         const val EXTRA_SELECTED_VIDEOS = "selected_videos"
         const val EXTRA_DEFAULT_PAGE = "extra_default_page"
     }
 
+    private var _baseBinding: OssFragmentBaseBinding? = null
+    val baseBinding get() = _baseBinding!!
+
+    lateinit var childView: View
+
     @Suppress("UNCHECKED_CAST")
-    protected fun getPhotosFromArguments() : List<PhotoFile>{
+    protected fun getPhotosFromArguments(): List<PhotoFile> {
         this.arguments?.let {
-            if(it.containsKey(EXTRA_SELECTED_PHOTOS)){
+            if (it.containsKey(EXTRA_SELECTED_PHOTOS)) {
                 return it.getSerializable(EXTRA_SELECTED_PHOTOS) as List<PhotoFile>
             }
         }
@@ -34,19 +37,24 @@ open abstract class BaseFragment : Fragment() {
     }
 
     @Suppress("UNCHECKED_CAST")
-    protected fun getVideosFromArguments() : List<VideoFile>{
-       this.arguments?.let {
-            if(it.containsKey(EXTRA_SELECTED_VIDEOS)){
+    protected fun getVideosFromArguments(): List<VideoFile> {
+        this.arguments?.let {
+            if (it.containsKey(EXTRA_SELECTED_VIDEOS)) {
                 return it.getSerializable(EXTRA_SELECTED_VIDEOS) as List<VideoFile>
             }
         }
-        return  emptyList()
+        return emptyList()
     }
 
-    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
-        return inflater.inflate(R.layout.oss_fragment_base, container, false).apply {
-            baseContainer.addView(inflater.inflate(getLayoutId(),null))
-        }
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        _baseBinding = OssFragmentBaseBinding.inflate(layoutInflater, container, false)
+        childView = inflater.inflate(getLayoutId(),null)
+        baseBinding.baseContainer.addView(childView)
+        return baseBinding.root
     }
 
     @LayoutRes
@@ -61,55 +69,58 @@ open abstract class BaseFragment : Fragment() {
 
     @CallSuper
     private fun setToolbar() {
-        toolbarTitle.text = getScreenTitle()
-        toolbarTitle.setTextColor(context!!.resources!!.getColor(R.color.oss_toolbar_text))
-        if(setHomeAsUp()){
-            toolbarBackButton.visibility = View.VISIBLE
-            toolbarBackButton.setImageResource(getHomeAsUpIcon())
-        }else{
-            toolbarBackButton.visibility = View.GONE
+        baseBinding.customToolbar.apply {
+            toolbarTitle.text = getScreenTitle()
+            toolbarTitle.setTextColor(requireContext().resources.getColor(R.color.oss_toolbar_text))
+            if (setHomeAsUp()) {
+                toolbarBackButton.visibility = View.VISIBLE
+                toolbarBackButton.setImageResource(getHomeAsUpIcon())
+            } else {
+                toolbarBackButton.visibility = View.GONE
+            }
+            if (shouldHideToolBar()) {
+                toolbarView.visibility = View.GONE
+            }
+            toolbarBackButton.setOnClickListener { onBackPressed() }
         }
-        if(shouldHideToolBar()){
-            toolbarView.visibility = View.GONE
-        }
-        toolbarBackButton.setOnClickListener { onBackPressed() }
     }
 
     override fun onOptionsItemSelected(item: MenuItem): Boolean {
-        return if(item.itemId == android.R.id.home){
+        return if (item.itemId == android.R.id.home) {
             onBackPressed()
             return true
-        }else{
+        } else {
             super.onOptionsItemSelected(item)
         }
 
     }
 
     protected fun showToolbar() {
-        toolbarView.visibility = View.VISIBLE
+        baseBinding.customToolbar.toolbarView.visibility = View.VISIBLE
     }
 
     protected fun hideToolbar() {
-        toolbarView.visibility = View.GONE
+        baseBinding.customToolbar.toolbarView.visibility = View.GONE
     }
 
     abstract fun onBackPressed()
 
-    open fun getHomeAsUpIcon()  = R.drawable.oss_media_ic_back
+    open fun getHomeAsUpIcon() = R.drawable.oss_media_ic_back
 
-    open fun setHomeAsUp()  = false
+    open fun setHomeAsUp() = false
 
-    abstract fun getScreenTitle() : String
+    abstract fun getScreenTitle(): String
 
-    open fun shouldHideToolBar()  = false
+    open fun shouldHideToolBar() = false
 
     abstract fun setUpViews()
 
     @CallSuper
-    open fun initViewModels(){}
+    open fun initViewModels() {
+    }
 
     override fun onDestroyView() {
         super.onDestroyView()
-        baseToolbar.visibility = View.VISIBLE
+        baseBinding.baseToolbar.visibility = View.VISIBLE
     }
 }
