@@ -2,10 +2,8 @@ package com.mediapicker.gallery.presentation.fragments
 
 import android.app.Activity
 import com.mediapicker.gallery.presentation.activity.GalleryActionListener
-import kotlinx.android.synthetic.main.oss_fragment_folder_view.*
 
 abstract class BaseGalleryViewFragment : BaseFragment() {
-
 
     protected var galleryActionListener: GalleryActionListener? = null
 
@@ -16,14 +14,11 @@ abstract class BaseGalleryViewFragment : BaseFragment() {
         }
     }
 
-    override fun setUpViews() {
-        actionButton.setOnClickListener { onActionButtonClick() }
-    }
+    override fun setUpViews() {}
 
     open fun onActionButtonClick() {}
 
     override fun onBackPressed() {
         activity?.onBackPressed()
     }
-
 }
