@@ -1,6 +1,7 @@
 package com.mediapicker.gallery.presentation.fragments
 
 import android.view.View
+import android.widget.ProgressBar
 import androidx.lifecycle.Observer
 import androidx.recyclerview.widget.GridLayoutManager
 import androidx.recyclerview.widget.RecyclerView
@@ -13,24 +14,30 @@ import com.mediapicker.gallery.presentation.viewmodels.StateData
 import com.mediapicker.gallery.presentation.viewmodels.factory.BaseLoadMediaViewModel
 import com.mediapicker.gallery.util.ItemOffsetDecoration
 import com.mediapicker.gallery.utils.SnackbarUtils
-import kotlinx.android.synthetic.main.oss_fragment_gallery.*
 
 abstract class BaseViewPagerItemFragment : BaseFragment() {
 
     var pageTitle = ""
 
-    protected val bridgeViewModel : BridgeViewModel by lazy {
-        getActivityScopedViewModel{ BridgeViewModel(emptyList(), emptyList(),Gallery.galleryConfig) }
+    protected val bridgeViewModel: BridgeViewModel by lazy {
+        getActivityScopedViewModel {
+            BridgeViewModel(
+                emptyList(),
+                emptyList(),
+                Gallery.galleryConfig
+            )
+        }
     }
 
     override fun initViewModels() {
         super.initViewModels()
         bridgeViewModel.getMediaStateLiveData().observe(this, Observer { reloadMedia() })
-        getBaseLoadMediaViewModel().getLoadingState().observe(this, Observer { handleLoadingState(it) })
+        getBaseLoadMediaViewModel().getLoadingState()
+            .observe(this, Observer { handleLoadingState(it) })
     }
 
     override fun setUpViews() {
-        ossRecycleView.apply {
+        childView.findViewById<RecyclerView>(R.id.ossRecycleView).apply {
             val spacing = resources.getDimensionPixelSize(R.dimen.gallery_item_offset)
             val gridLayoutManager = GridLayoutManager(context, 3)
             this.layoutManager = gridLayoutManager
@@ -45,9 +52,9 @@ abstract class BaseViewPagerItemFragment : BaseFragment() {
 
     override fun getLayoutId() = R.layout.oss_fragment_gallery
 
-    abstract fun getBaseLoadMediaViewModel() : BaseLoadMediaViewModel
+    abstract fun getBaseLoadMediaViewModel(): BaseLoadMediaViewModel
 
-    protected open fun reloadMedia(){
+    protected open fun reloadMedia() {
         getBaseLoadMediaViewModel().loadMedia(this)
     }
 
@@ -55,20 +62,23 @@ abstract class BaseViewPagerItemFragment : BaseFragment() {
         when (stateData) {
             StateData.SUCCESS -> hideProgressBar()
             StateData.LOADING -> showProgressBar()
+            StateData.ERROR -> {
+                // no-op
+            }
         }
     }
 
     protected open fun hideProgressBar() {
-        progressBar.visibility = View.GONE
-        ossRecycleView.visibility = View.VISIBLE
+        childView.findViewById<ProgressBar>(R.id.progressBar).visibility = View.GONE
+        childView.findViewById<RecyclerView>(R.id.ossRecycleView).visibility = View.VISIBLE
     }
 
     protected open fun showProgressBar() {
-        progressBar.visibility = View.VISIBLE
-        ossRecycleView.visibility = View.GONE
+        childView.findViewById<ProgressBar>(R.id.progressBar).visibility = View.VISIBLE
+        childView.findViewById<RecyclerView>(R.id.ossRecycleView).visibility = View.GONE
     }
 
-    protected open fun showMsg(msg : String){
+    protected open fun showMsg(msg: String) {
         SnackbarUtils.show(view, msg, Snackbar.LENGTH_LONG)
     }
 
