@@ -1,8 +1,9 @@
 package com.mediapicker.gallery.presentation.fragments
 
-import android.util.Log
+import androidx.appcompat.widget.AppCompatButton
 import androidx.lifecycle.Observer
 import androidx.recyclerview.widget.GridLayoutManager
+import androidx.recyclerview.widget.RecyclerView
 import com.mediapicker.gallery.Gallery
 import com.mediapicker.gallery.R
 import com.mediapicker.gallery.data.repositories.GalleryService
@@ -12,12 +13,8 @@ import com.mediapicker.gallery.presentation.adapters.GalleryFolderAdapter
 import com.mediapicker.gallery.presentation.utils.ItemDecorationAlbumColumns
 import com.mediapicker.gallery.presentation.utils.getFragmentScopedViewModel
 import com.mediapicker.gallery.presentation.viewmodels.LoadAlbumViewModel
-import kotlinx.android.synthetic.main.oss_custom_toolbar.*
-import kotlinx.android.synthetic.main.oss_fragment_carousal.*
-import kotlinx.android.synthetic.main.oss_fragment_folder_view.*
 
 class FolderViewFragment : BaseGalleryViewFragment(), OnItemClickListener<PhotoAlbum> {
-
 
     private val loadAlbumViewModel: LoadAlbumViewModel by lazy {
         getFragmentScopedViewModel { LoadAlbumViewModel(GalleryService(Gallery.getApp())) }
@@ -26,10 +23,12 @@ class FolderViewFragment : BaseGalleryViewFragment(), OnItemClickListener<PhotoA
     override fun getScreenTitle() = getString(R.string.oss_title_folder_fragment)
 
     private lateinit var adapter: GalleryFolderAdapter
+    private lateinit var folderRV: RecyclerView
+    private lateinit var actionButton: AppCompatButton
 
     private fun setAlbumData(setOfAlbum: HashSet<PhotoAlbum>) {
         adapter.apply {
-            var albumList=mutableListOf<PhotoAlbum>().apply { this.addAll(setOfAlbum) }
+            var albumList = mutableListOf<PhotoAlbum>().apply { this.addAll(setOfAlbum) }
             this.listOfFolders = albumList.sortedBy { it.name }
             notifyDataSetChanged()
         }
@@ -39,22 +38,40 @@ class FolderViewFragment : BaseGalleryViewFragment(), OnItemClickListener<PhotoA
 
     override fun setUpViews() {
         super.setUpViews()
-        adapter = GalleryFolderAdapter(context!!, listOfFolders = emptyList(), onItemClickListener = this)
+        adapter = GalleryFolderAdapter(
+            requireContext(),
+            listOfFolders = emptyList(),
+            onItemClickListener = this
+        )
+
+        folderRV = childView.findViewById(R.id.folderRV)
+        actionButton = childView.findViewById(R.id.actionButton)
+
         folderRV.apply {
-            this.addItemDecoration(ItemDecorationAlbumColumns(resources.getDimensionPixelSize(R.dimen.module_base), COLUMNS_COUNT))
+            this.addItemDecoration(
+                ItemDecorationAlbumColumns(
+                    resources.getDimensionPixelSize(R.dimen.module_base),
+                    COLUMNS_COUNT
+                )
+            )
             this.layoutManager = GridLayoutManager(this@FolderViewFragment.activity, COLUMNS_COUNT)
             this.adapter = this@FolderViewFragment.adapter
         }
+
         actionButton.isSelected = true
 
         if (Gallery.galleryConfig.galleryLabels.galleryFolderAction.isNotBlank()) {
             actionButton.text = Gallery.galleryConfig.galleryLabels.galleryFolderAction
         }
 
-        toolbarTitle.isAllCaps = Gallery.galleryConfig.textAllCaps
         actionButton.isAllCaps = Gallery.galleryConfig.textAllCaps
+        actionButton.setOnClickListener { onActionButtonClick() }
 
-        toolbarTitle.gravity = Gallery.galleryConfig.galleryLabels.titleAlignment
+
+        baseBinding.customToolbar.apply {
+            toolbarTitle.isAllCaps = Gallery.galleryConfig.textAllCaps
+            toolbarTitle.gravity = Gallery.galleryConfig.galleryLabels.titleAlignment
+        }
     }
 
     override fun initViewModels() {
